refactor(perl-dap): split dispatcher into SRP submodules (#0000)
